refactor: extract TaxPayersPanel and TaxReceiptsPanel into reusable components
