+++ src/components/tax-payers-panel/index.ts
@@ -1,0 +1,1 @@
+export { default as TaxPayersPanel } from './tax-payers-panel';
