+++ src/components/tax-receipts-panel/index.ts
@@ -1,0 +1,1 @@
+export { default as TaxReceiptsPanel } from './tax-receipts-panel';
