+++ src/pages/App.tsx
@@ -1,37 +1,9 @@
 import { useState } from 'react';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
-import { Skeleton } from '@/components/ui/skeleton';
-import {
-  ChevronRight,
-  ArrowLeft,
-  CircleHelp,
-  User,
-  Building2,
-} from 'lucide-react';
 
-import {
-  getDisplayName,
-  getDocLabel,
-  getStatus,
-  formatRD,
-  getDocLabelFromTaxPayer,
-} from '@/utils';
-import {
-  TAX_PAYER_TYPES,
-  type TaxPayer,
-  type TaxPayerTypeId,
-} from '@/types/tax-payer';
+import { TaxPayersPanel } from '@/components/tax-payers-panel';
+import { TaxReceiptsPanel } from '@/components/tax-receipts-panel';
 import { useTaxPayer, useTaxReceipts } from '@/services/queries/tax-payer';
-import { QueryError } from '@/components/common/query-error';
+import { type TaxPayerTypeId } from '@/types/tax-payer';
 
 export default function RegistryPage() {
   const [selectedId, setSelectedId] = useState<number | null>(null);
@@ -58,10 +30,6 @@ export default function RegistryPage() {
     taxPayerId: selectedId ?? 0,
   });
 
-  const selectedRegistry = taxPayers?.data.find(
-    (taxPayer: TaxPayer) => taxPayer.id === selectedId
-  );
-
   return (
     <div className="container mx-auto p-6 max-w-6xl">
       <div className="mb-8">
@@ -72,252 +40,26 @@ export default function RegistryPage() {
       </div>
 
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-        <div className="space-y-4">
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                {Boolean(selectedId) && (
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onClick={() => setSelectedId(0)}
-                    className="p-1 h-auto"
-                  >
-                    <ArrowLeft className="h-4 w-4" />
-                  </Button>
-                )}
-                Contribuyentes
-              </CardTitle>
-            </CardHeader>
-            <CardContent className="space-y-4">
-              <div className="flex flex-wrap items-center gap-2">
-                {TAX_PAYER_TYPES.map((t) => {
-                  const selected = taxPayerTypeId === t.id;
-                  return (
-                    <Button
-                      key={t.id}
-                      type="button"
-                      size="sm"
-                      variant={selected ? 'default' : 'outline'}
-                      aria-pressed={selected}
-                      onClick={() => {
-                        setTaxPayerTypeId((prev) =>
-                          prev === t.id ? null : t.id
-                        );
-                        setSelectedId(null);
-                      }}
-                      className="flex items-center gap-2"
-                    >
-                      {t.code === 'PER' ? (
-                        <User className="h-4 w-4" />
-                      ) : (
-                        <Building2 className="h-4 w-4" />
-                      )}
-                      {t.description}
-                    </Button>
-                  );
-                })}
-              </div>
-              <div className="p-1 space-y-2 max-h-96 overflow-y-auto">
-                {isErrorTaxPayer ? (
-                  <QueryError
-                    error={taxPayersError}
-                    onRetry={refetchTaxPayers}
-                    className="my-2"
-                  />
-                ) : isLoadingTaxPayers ? (
-                  Array.from({ length: 3 }).map((_, i) => (
-                    <Card key={i} className="p-4">
-                      <div className="space-y-2">
-                        <Skeleton className="h-4 w-3/4" />
-                        <Skeleton className="h-3 w-full" />
-                        <Skeleton className="h-3 w-1/4" />
-                      </div>
-                    </Card>
-                  ))
-                ) : taxPayers?.data?.length === 0 ? (
-                  <div className="text-center py-8 text-muted-foreground">
-                    No se encontraron registros
-                  </div>
-                ) : (
-                  taxPayers?.data?.map((taxPayer: TaxPayer) => {
-                    const name = getDisplayName(taxPayer);
-                    const doc = getDocLabel(taxPayer);
-                    const tipo = taxPayer?.taxPayerType?.description ?? '—';
-                    const estado = getStatus(taxPayer);
-
-                    return (
-                      <Card
-                        key={taxPayer.id}
-                        className={`cursor-pointer transition-colors hover:bg-accent/60 ${
-                          selectedId === taxPayer.id
-                            ? 'ring-2 ring-primary'
-                            : ''
-                        }`}
-                        onClick={() => setSelectedId(taxPayer.id)}
-                      >
-                        <CardContent className="p-4">
-                          <div className="flex items-start justify-between gap-3">
-                            <div className="flex-1 min-w-0">
-                              <h3 className="font-medium leading-tight truncate">
-                                {name}
-                              </h3>
-                              <p className="text-sm text-muted-foreground mt-1 truncate">
-                                {doc}
-                              </p>
-
-                              <div className="flex flex-wrap items-center gap-2 mt-2">
-                                <Badge variant="secondary">{tipo}</Badge>
-                                <Badge
-                                  variant={
-                                    estado === 'Activo'
-                                      ? 'outline'
-                                      : 'destructive'
-                                  }
-                                >
-                                  {estado}
-                                </Badge>
-                              </div>
-                            </div>
-
-                            <ChevronRight className="h-4 w-4 shrink-0 text-muted-foreground mt-1" />
-                          </div>
-                        </CardContent>
-                      </Card>
-                    );
-                  })
-                )}
-              </div>
-            </CardContent>
-          </Card>
-        </div>
-
-        <div className="space-y-4">
-          <Card>
-            <CardHeader>
-              <CardTitle>Comprobantes Fiscales</CardTitle>
-              <CardDescription>
-                {selectedRegistry
-                  ? `Comprobantes Fiscales relacionados a ${getDisplayName(
-                      selectedRegistry
-                    )}`
-                  : ''}
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              {!selectedId ? (
-                <div
-                  role="status"
-                  aria-live="polite"
-                  className="text-center py-12 text-muted-foreground"
-                >
-                  <CircleHelp className="h-12 w-12 mx-auto opacity-50 mb-4" />
-                  <p className="text-sm">
-                    Seleccione un contribuyente{' '}
-                    <span className="hidden md:inline">
-                      en el panel izquierdo{' '}
-                    </span>
-                    para visualizar sus comprobantes fiscales
-                  </p>
-                </div>
-              ) : isErrorTaxReceipts ? (
-                <QueryError
-                  error={taxReceiptsError}
-                  onRetry={refetchTaxReceipts}
-                  className="my-2"
-                />
-              ) : isLoadingTaxReceipts || isFetchingTaxRecipts ? (
-                <div className="p-1 space-y-3">
-                  {Array.from({ length: 3 }).map((_, i) => (
-                    <Card key={i} className="p-4">
-                      <div className="space-y-2">
-                        <Skeleton className="h-4 w-2/3" />
-                        <Skeleton className="h-3 w-full" />
-                        <Skeleton className="h-6 w-16" />
-                      </div>
-                      <div></div>
-                    </Card>
-                  ))}
-                </div>
-              ) : taxReceipts?.data?.length === 0 ? (
-                <div className="text-center py-8 text-muted-foreground">
-                  No related registries found
-                </div>
-              ) : (
-                <div className="space-y-3 max-h-96 overflow-y-auto">
-                  {taxReceipts?.data?.map((receipt) => {
-                    const name = getDisplayName(receipt.taxPayer);
-                    const docLabel = getDocLabelFromTaxPayer(receipt.taxPayer);
-                    const tipo =
-                      receipt.taxPayer.taxPayerType.description ?? '—';
-                    const docTipo =
-                      receipt.taxPayer.documentType.description ?? '—';
-
-                    return (
-                      <Card
-                        key={receipt.id}
-                        className="hover:bg-accent transition-colors"
-                      >
-                        <CardContent className="p-4">
-                          <div className="flex items-start justify-between gap-4">
-                            <div className="min-w-0">
-                              <h4 className="font-medium leading-tight truncate">
-                                NCF:{' '}
-                                <span className="font-mono">{receipt.ncf}</span>
-                              </h4>
-                              <p className="text-sm text-muted-foreground mt-1 truncate">
-                                {name}
-                              </p>
-                              <p className="text-xs text-muted-foreground mt-0.5 truncate">
-                                {docLabel}
-                              </p>
-
-                              <div className="flex flex-wrap gap-2 mt-2">
-                                <Badge variant="secondary">{tipo}</Badge>
-                                <Badge variant="outline">{docTipo}</Badge>
-                              </div>
-                            </div>
-
-                            <div className="shrink-0 text-right">
-                              <div className="text-xs text-muted-foreground">
-                                Monto
-                              </div>
-                              <div className="font-semibold">
-                                {formatRD(receipt.amount)}
-                              </div>
-                              <div className="text-xs text-muted-foreground mt-2">
-                                ITBIS
-                              </div>
-                              <div className="font-bold text-primary">
-                                {formatRD(receipt.itbis)}
-                              </div>
-                            </div>
-                          </div>
-                        </CardContent>
-                      </Card>
-                    );
-                  })}
-                </div>
-              )}
-            </CardContent>
-            {Boolean(selectedId) &&
-              !isLoadingTaxReceipts &&
-              !isFetchingTaxRecipts &&
-              !isErrorTaxReceipts &&
-              (taxReceipts?.data?.length ?? 0) > 0 && (
-                <CardFooter className="flex justify-end">
-                  <span className="px-3 py-1 rounded-md bg-muted font-semibold text-primary">
-                    {`Total ITBIS: ${formatRD(
-                      taxReceipts!.data!.reduce(
-                        (sum, it) => sum + (it.itbis ?? 0),
-                        0
-                      )
-                    )}`}
-                  </span>
-                </CardFooter>
-              )}
-          </Card>
-        </div>
+        <TaxPayersPanel
+          selectedId={selectedId}
+          setSelectedId={setSelectedId}
+          taxPayerTypeId={taxPayerTypeId}
+          setTaxPayerTypeId={setTaxPayerTypeId}
+          taxPayers={taxPayers}
+          isLoadingTaxPayers={isLoadingTaxPayers}
+          isErrorTaxPayer={isErrorTaxPayer}
+          taxPayersError={taxPayersError}
+          refetchTaxPayers={refetchTaxPayers}
+        />
+        <TaxReceiptsPanel
+          selectedId={selectedId}
+          isLoadingTaxReceipts={isLoadingTaxReceipts}
+          isFetchingTaxReceipts={isFetchingTaxRecipts}
+          isErrorTaxReceipts={isErrorTaxReceipts}
+          taxReceipts={taxReceipts}
+          taxReceiptsError={taxReceiptsError}
+          refetchTaxReceipts={refetchTaxReceipts}
+        />
       </div>
     </div>
   );
